fix: share waitFor lifecycle cutoff
